SQLAgent/requirements.txt
# Core AI and Language Processing
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
google-generativeai>=0.5.4

# Database and Data Processing
SQLAlchemy>=2.0.0
pandas>=2.2.2
pydantic>=2.0.0

# Data Visualization
matplotlib>=3.9.0
seaborn>=0.12.0

# GUI and UI
tkinter-tooltip>=2.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Additional Dependencies for Material Design
Pillow>=10.0.0  # For image processing in charts
numpy>=1.24.0   # For numerical operations
